Remove EntityManagerProvider
