+++ cli-config.php
@@ -2,16 +2,12 @@
 
 use Doctrine\DBAL\DriverManager;
 use Doctrine\ORM\Tools\Console\ConsoleRunner;
-use WMDE\Fundraising\Store\EntityManagerProvider;
+use WMDE\Fundraising\Store\Factory;
 
 require_once __DIR__ . '/vendor/autoload.php';
 
 $config = parse_ini_file( __DIR__ . '/config.ini', true );
 
-$provider = new EntityManagerProvider( DriverManager::getConnection( $config['database'] ) );
-$entityManager = $provider->getEntityManager();
+$factory = new Factory( DriverManager::getConnection( $config['database'] ) );
 
-$platform = $entityManager->getConnection()->getDatabasePlatform();
-$platform->registerDoctrineTypeMapping( 'enum', 'string' );
-
-return ConsoleRunner::createHelperSet( $entityManager );
+return ConsoleRunner::createHelperSet( $factory->getEntityManager() );
